--- a/src/clojush/util.clj
+++ b/src/clojush/util.clj
@@ -2,8 +2,8 @@
   (:use clojush.globals)
   (:require [clojure.math.numeric-tower :as math]
             [clojure.zip :as zip]
-            [clojure.walk :as walk]
-            [clojure.string :as string]))
+            [clojure.string :as string])
+  (:import [info.debatty.java.stringsimilarity Levenshtein]))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; utilities
@@ -19,7 +19,7 @@
      :vector_float (fn [thing] (and (vector? thing) (float? (first thing))))
      :vector_string (fn [thing] (and (vector? thing) (string? (first thing))))
      :vector_boolean (fn [thing] (and (vector? thing) (or (= (first thing) true) (= (first thing) false))))}))
-     
+
 (defn recognize-literal
   "If thing is a literal, return its type -- otherwise return false."
   [thing]
@@ -99,43 +99,43 @@
   [tree]
   (loop [remaining tree
          total 0]
-    (cond (not (seq? remaining)) 
+    (cond (not (seq? remaining))
           total
-          ;; 
-          (empty? remaining) 
+          ;;
+          (empty? remaining)
           (inc total)
           ;;
-          (not (seq? (first remaining))) 
-          (recur (rest remaining) 
+          (not (seq? (first remaining)))
+          (recur (rest remaining)
                  total)
           ;;
-          :else 
-          (recur (list-concat (first remaining) 
-                              (rest remaining)) 
+          :else
+          (recur (list-concat (first remaining)
+                              (rest remaining))
                  (inc total)))))
 
 (defn count-points
-  "Returns the number of points in tree, where each atom and each pair of parentheses 
+  "Returns the number of points in tree, where each atom and each pair of parentheses
    counts as a point."
   [tree]
   (loop [remaining tree
          total 0]
-    (cond (not (seq? remaining)) 
-          (inc total)
-          ;; 
-          (empty? remaining) 
+    (cond (not (seq? remaining))
           (inc total)
           ;;
-          (not (seq? (first remaining))) 
-          (recur (rest remaining) 
+          (empty? remaining)
+          (inc total)
+          ;;
+          (not (seq? (first remaining)))
+          (recur (rest remaining)
                  (inc total))
           ;;
-          :else 
-          (recur (list-concat (first remaining) 
-                              (rest remaining)) 
+          :else
+          (recur (list-concat (first remaining)
+                              (rest remaining))
                  (inc total)))))
 
-(defn code-at-point 
+(defn code-at-point
   "Returns a subtree of tree indexed by point-index in a depth first traversal."
   [tree point-index]
   (let [index (mod (math/abs point-index) (count-points tree))
@@ -177,7 +177,7 @@
 
 ; Note: Well, I (Tom) think I figured out why truncate was there. When I tried running
 ; the change problem, it threw an exception trying to cast into an int a number
-; that was too big. Maybe there's a different principled way to use casting, but 
+; that was too big. Maybe there's a different principled way to use casting, but
 ; I'm just going to add truncate back for now!
 (defn truncate
   "Returns a truncated integer version of n."
@@ -213,7 +213,7 @@
   (postwalklist (fn [x] (if (contains? smap x) (smap x) x)) form))
 
 (defn subst
-  "Returns the given list but with all instances of that (at any depth)                                   
+  "Returns the given list but with all instances of that (at any depth)
    replaced with this. Read as 'subst this for that in list'. "
   [this that lst]
   (postwalklist-replace {that this} lst))
@@ -222,7 +222,7 @@
   "Returns true if tree contains subtree at any level. Inefficient but
    functional implementation."
   [tree subtree]
-  (or 
+  (or
     (= tree subtree)
     (not (= tree (subst (gensym) subtree tree)))))
 
@@ -232,7 +232,7 @@
    subtree. For example, (contining-subtree '(b (c (a)) (d (a))) '(a)) => (c (a)).
    Returns nil if tree does not contain subtree."
   [tree subtree]
-  (cond 
+  (cond
     (not (seq? tree)) nil
     (empty? tree) nil
     (some #{subtree} tree) tree
@@ -283,9 +283,9 @@
    cases."
   [domains]
   (vec
-    (apply 
-      mapv 
-      concat 
+    (apply
+      mapv
+      concat
       (map (fn [[input-set n-train n-test]]
              (if (fn? input-set)
                (vector (repeatedly n-train input-set)
@@ -297,66 +297,19 @@
                      test-inputs (if (= n-test (count input-set))
                                    input-set ; NOTE: input-set is not shuffled if the same size as n-test
                                    (drop n-train shuffled-inputs))]
-                 (assert (= (+ n-train n-test) (count input-set)) 
+                 (assert (= (+ n-train n-test) (count input-set))
                          "Sizes of train and test sets don't add up to the size of the input set.")
                  (vector train-inputs test-inputs))))
            domains))))
 
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; from https://github.com/KushalP/mailcheck-clj/blob/master/src/mailcheck/levenshtein.clj
-
-(defn compute-next-row
-  "computes the next row using the prev-row current-element and the other seq"
-  [prev-row current-element other-seq pred]
-  (reduce
-    (fn [row [diagonal above other-element]]
-      (let [update-val (if (pred other-element current-element)
-                         ;; if the elements are deemed equivalent according to the predicate
-                         ;; pred, then no change has taken place to the string, so we are
-                         ;; going to set it the same value as diagonal (which is the previous edit-distance)
-                         diagonal
-                         ;; in the case where the elements are not considered equivalent, then we are going
-                         ;; to figure out if its a substitution (then there is a change of 1 from the previous
-                         ;; edit distance) thus the value is diagonal + 1 or if its a deletion, then the value
-                         ;; is present in the columns, but not in the rows, the edit distance is the edit-distance
-                         ;; of last of row + 1 (since we will be using vectors, peek is more efficient)
-                         ;; or it could be a case of insertion, then the value is above+1, and we chose
-                         ;; the minimum of the three
-                         (inc (min diagonal above (peek row))))]
-                         
-        (conj row update-val)))
-    ;; we need to initialize the reduce function with the value of a row, since we are
-    ;; constructing this row from the previous one, the row is a vector of 1 element which
-    ;; consists of 1 + the first element in the previous row (edit distance between the prefix so far
-    ;; and an empty string)
-    [(inc (first prev-row))]
-    ;; for the reduction to go over, we need to provide it with three values, the diagonal
-    ;; which is the same as prev-row because it starts from 0, the above, which is the next element
-    ;; from the list and finally the element from the other sequence itself.
-    (map vector prev-row (next prev-row) other-seq)))
+(def l (new Levenshtein))
 
 (defn levenshtein-distance
   "Levenshtein Distance - http://en.wikipedia.org/wiki/Levenshtein_distance
      In information theory and computer science, the Levenshtein distance is a
-     metric for measuring the amount of difference between two sequences. This
-     is a functional implementation of the levenshtein edit
-     distance with as little mutability as possible.
-     Still maintains the O(n*m) guarantee."
-  [a b & {p :predicate  :or {p =}}]
-  (cond
-    (empty? a) (count b)
-    (empty? b) (count a)
-    :else (peek
-            (reduce
-              ;; we use a simple reduction to convert the previous row into the next-row  using the
-              ;; compute-next-row which takes a current element, the previous-row computed so far
-              ;; and the predicate to compare for equality.
-              (fn [prev-row current-element]
-                (compute-next-row prev-row current-element b p))
-              ;; we need to initialize the prev-row with the edit distance between the various prefixes of
-              ;; b and the empty string.
-              (range (inc (count b)))
-              a))))
+     metric for measuring the amount of difference between two sequences."
+  [a b]
+  (.distance l a b))
 
 (defn sequence-similarity
   [sequence1 sequence2]
@@ -374,7 +327,7 @@
   "Calculates the Hamming distance between two sequences, including strings"
   [seq1 seq2]
   (apply + (map #(if (= %1 %2) 0 1)
-                  seq1 seq2)))
+                 seq1 seq2)))
 
 ;;;;;;;;;;;;;;:::::;;;;;;;;;;;;;;
 ;; Simple Statistic Functions