TODO:

- [ ] Change benchmarks to another framework that is more reliable

* Use protocol for figuring out how to deal with item in exec stack
  instead of case matching
* replace string distance with java version that is faster
* cache potential children for lexicase selection

## New bencmharks

```
Measuring clojush.bench.core-bench

main-autocon-10-gen-jan-13 (core_bench.clj:47)

time: 204.044867 sec, sd: 27.915311 sec

main-autocon-10-gen-oct-12 (core_bench.clj:50)

time: 236.894546 sec, sd: 2.198388 sec

main-autocon-10-gen-oct-12-serial (core_bench.clj:53)

time: 1447.143918 sec, sd: 34.121085 sec

main-nth-prime-10-gen (core_bench.clj:56)

time: 69.681388 sec, sd: 231.904642 ms

cleanup (core_bench.clj:60)

Measuring clojush.bench.interpreter-bench

eval-push-on-10000-from-clojush.problems.software.replace-space-with-newline (interpreter_bench.clj:55)
Grabbing inputs...
Running benchmark...

time: 14.379836 sec, sd: 944.144724 µs
```

## Old benchmarks

```
aul@deucalion:~/Clojush$ lein trampoline libra

Measuring clojush.bench.core-bench

main-autocon-10-gen-jan-13 (core_bench.clj:47)

time: 244.811248 sec, sd: 1.451128 sec

main-autocon-10-gen-oct-12 (core_bench.clj:50)

time: 303.604710 sec, sd: 1.185334 sec

main-autocon-10-gen-oct-12-serial (core_bench.clj:53)

time: 2771.459385 sec, sd: 64.543268 sec

main-nth-prime-10-gen (core_bench.clj:56)

time: 197.568371 sec, sd: 229.763320 ms

cleanup (core_bench.clj:60)

Measuring clojush.bench.interpreter-bench

eval-push-on-10000-from-clojush.problems.software.replace-space-with-newline (interpreter_bench.clj:50)
Grabbing inputs...
Running benchmark...

time: 20.113981 sec, sd: 1.259571 ms

```


## String distance timings
```
Measuring clojush.bench.util-bench

levenshtein-distance-on-10000-from-clojush.problems.software.replace-space-with-newline (util_bench.clj:14)
Grabbing inputs...
Running benchmark...

time: 9.789159 ms, sd: 320.935487 ns
Running alternative... #object[clj_fuzzy.levenshtein$distance 0x6ec3d8e4 clj_fuzzy.levenshtein$distance@6ec3d8e4]

time: 653.200526 ms, sd: 1.890006 ms
Running alternative... #object[clojush.bench.util_bench$fn__5861$fn__5876 0x101c15ad clojush.bench.util_bench$fn__5861$fn__5876@101c15ad]

time: 10.146529 ms, sd: 53.602368 ns
```